requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
